src/main/java/com/galvanize/heroapp/service/HeroService.java
package com.galvanize.heroapp.service;

import com.galvanize.heroapp.model.HeroResponse;
import com.galvanize.heroapp.repository.HeroRepository;
import org.springframework.stereotype.Service;

import java.util.ArrayList;
import java.util.List;

@Service
public class HeroService {

    private List<String> names;

    private HeroRepository heroRepository;

    public HeroService(HeroRepository heroRepository) {
        this.heroRepository = heroRepository;
    }

    public List<HeroResponse> getAllHeroNames() {
        return heroRepository.findAll();
    }
}
